try to fix the windows build
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@3165 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ tsan/ts_trace_info.h
@@ -66,10 +66,10 @@ struct MopInfo {
   bool      create_sblock() { return create_sblock_; }
 
  private:
-  uint64_t  pc_            :58;  // 48 bits is enough for pc, even on x86-64.
-  uintptr_t  create_sblock_ :1;
-  uintptr_t  is_write_      :1;
-  uintptr_t  size_minus1_   :4;  // 0..15
+  uint64_t  pc_           :58;  // 48 bits is enough for pc, even on x86-64.
+  uint64_t  create_sblock_ :1;
+  uint64_t  is_write_      :1;
+  uint64_t  size_minus1_   :4;  // 0..15
 };
 
 // ---------------- Lite Race ------------------
